some more fixes

+++ marketplace/app_shops/templatetags/discount.py
@@ -5,4 +5,4 @@ register = template.Library()
 
 @register.simple_tag()
 def apply_discount(value, discount):
-    return round(int(value) * 0.01 * (100 - discount), 2)
+    return round(float(value) * 0.01 * (100 - discount), 2)
